Add solution for problem `A - Skyscrapers` Link: https://codeforces.com/contest/1137/problem/A Tags: implementation, sortings, *1600 Ref: https://codeforces.com/contest/1137/submission/133964618

+++ codeforces/1137/A.cpp
@@ -7,7 +7,7 @@ using namespace std;
 #define dbg(...)
 #endif
 using ll = int64_t;
-auto sz = [](const auto &container) -> int { return int(container.size()); };
+auto sz = [](const auto &container) -> ll { return container.size(); };
 
 #pragma GCC optimize("O3,unroll-loops")
 #pragma GCC target("avx2,bmi,bmi2,lzcnt,popcnt")
